==> 02_luther/notebooks/03_goodreads_api.ipynb <==
### Goal: use the Goodreads API to gather data about books and authors on the _New York Times_ Fiction Best Sellers list for a range of years between 1942-2016

* General Search
* Get data by author id
* Get data by book id

import requests
import urllib

from bs4 import BeautifulSoup
import xml.etree.ElementTree as ET
import xml.dom.minidom
from lxml import etree

from time import sleep

import re

import pandas as pd
import numpy as np

import pickle

#goodreads api key
key = '2Kp7IGlO7GKjmAASOT2pEw'

with open('../data/interim/books_dict.pkl', 'rb') as picklefile:
    books = pickle.load(picklefile)

#books['A Tree Grows in Brooklyn']

def try_url(url):
    '''
    Attempt to access Goodreads page
    '''
    response = requests.get(url)
    status = response.status_code
    if status != 200:
        print status
    else:
        page = response.text
        soup = BeautifulSoup(page, 'xml')
        return soup

### General Search
url = https://www.goodreads.com/search.xml?key=YOUR_KEY&q=Ender%27s+Game

Requires
* Title of book

Returns 
* first item in results>work
    * books_count: number of versions
    * id: Goodreads ID
    * original_publication_day: publication day
    * original_publication_month: publication month
    * original_publication_year: publication year
    * average_rating: average rating
    * ratings_count: number of ratings
    * text_reviews_count: number of text reviews
* first item in results>work>best_book>author
    * id: author ID

search_base_url = 'https://www.goodreads.com/search.xml?'
search_criteria = {'key': key,
                  'q': None}

#books = {'A Tree Grows in Brooklyn' : 'Betty  Smith'}

search_urls = {}

for nyt_title, author in books.items():
    search_criteria['q'] = nyt_title + ' ' + author
    url = search_base_url + urllib.urlencode(search_criteria)
    search_urls[url] = [author, nyt_title]

def get_search_data(works, author, nyt_title, df):
    '''
    Retrieve data from works xml tree from search results page
    '''
    for work in works:
        if work.find('best_book/author/name').text == author:
            b_id = work.find('best_book/id').text
            gr_title = work.find('best_book/title').text
            b_count = work.find('books_count').text
            pub_day = work.find('original_publication_day').text
            pub_mon = work.find('original_publication_month').text
            pub_yr = work.find('original_publication_year').text
            b_avg_rating = work.find('average_rating').text
            b_ratings_count = work.find('ratings_count').text
            b_txt_rev_count = work.find('text_reviews_count').text
            a_id = work.find('best_book/author/id').text
            results = pd.Series((b_id, gr_title, nyt_title, b_count, pub_day, pub_mon, pub_yr, b_avg_rating, b_ratings_count, b_txt_rev_count, a_id, author))
            df = df.append(results, ignore_index=True)
            break
    return df

test_url = 'https://www.goodreads.com/search.xml?q=%22R%22+Is+for+Ricochet+%2B+Sue+Grafton&key=2Kp7IGlO7GKjmAASOT2pEw'
soup = try_url(test_url)

lst = search_urls[test_url]
results = ET.fromstring(str(soup.find('results')))
works = results.findall('work')
author = lst[0].decode('utf-8')
nyt_title = lst[1]
print author, nyt_title
df = pd.DataFrame(data=None)
print get_search_data(works, author, nyt_title, df)

#get data by title, author from search API
books_df = pd.DataFrame(data=None)
count = 0

for url, lst in search_urls.items():
    soup = try_url(url)
    results = ET.fromstring(str(soup.find('results')))
    works = results.findall('work')
    author = lst[0].decode('utf-8')
    nyt_title = lst[1]
    books_df = get_search_data(works, author, nyt_title, books_df)
    count += 1
    print nyt_title, count
    sleep(1.5)

len(books_df)

columns = {0: 'b_id',
           1: 'gr_title',
           2: 'nyt_title',
           3: 'b_count',
           4: 'pub_day',
           5: 'pub_mon',
           6: 'pub_yr',
           7: 'b_avg_rating',
           8: 'b_ratings_count',
           9: 'b_txt_rev_count',
           10: 'a_id',
           11: 'author'}
books_df.rename(columns=columns, inplace=True)

Check which books did not return results from goodreads.

gr_titles = set(books_df['nyt_title'])
nyt_titles = set(books.keys())
missing_titles = nyt_titles.difference(gr_titles)

for title in missing_titles:
    print title, books[title]

len(missing_titles)

with open('../data/raw/books_scraped.pkl', 'wb') as picklefile:
    pickle.dump(books_df, picklefile)

### Get info about an author by id

url = https://www.goodreads.com/author/show/18541?format=xml&key=2Kp7IGlO7GKjmAASOT2pEw

Requires
* author ID

Returns
* in author
    * fans_count: number of fans on goodreads
    * works_count: number of works
    * gender: m/f
    * hometown: hometown
    * born_at: date of birth
    * died_at: date of death
 
_add these if necessary LATER_:
* in author>books>book[1]>authors>author[1]
    * average_rating: average rating for author
    * ratings_count: total ratings for author
    * text_reviews_count: total number of reviews for author

with open('../data/raw/books_scraped.pkl', 'rb') as picklefile:
    books_df = pickle.load(picklefile)

a_ids = books_df['a_id'].drop_duplicates()
len(a_ids)

def get_author_data(author, a_id, df):
    '''
    Retrieve data from author xml tree on author page
    '''
    a_fans_count = author.find('fans_count').text
    a_works_count = author.find('works_count').text
    gender = author.find('gender').text
    hometown = author.find('hometown').text
    birth_date = author.find('born_at').text
    death_date = author.find('died_at').text
    first_book = author.find('books').find('book').find('authors').find('author')
    a_avg_rating = first_book.find('average_rating').text
    a_ratings_count = first_book.find('ratings_count').text
    a_text_reviews_count = first_book.find('text_reviews_count').text
    results = pd.Series((a_id, a_fans_count, a_works_count, gender, hometown, birth_date, death_date, a_avg_rating, a_ratings_count, a_text_reviews_count))
    df = df.append(results, ignore_index=True)
    return df

author_base_url = 'https://www.goodreads.com/author/show/'
search_criteria = {'format': 'xml',
                  'key': key}

author_urls = {}

for a_id in a_ids:
    url = author_base_url + a_id + '?' + urllib.urlencode(search_criteria)
    author_urls[url] = a_id

#get data by author id from author API
authors_df = pd.DataFrame(data=None)
count = 0

for url, a_id in author_urls.items():
    soup = try_url(url)
    author = ET.fromstring(str(soup.find('author')))
    authors_df = get_author_data(author, a_id, authors_df)
    count += 1
    print a_id, count
    sleep(1.5)

columns = {0: 'a_id',
           1: 'a_fans_count',
           2: 'a_works_count',
           3: 'gender',
           4: 'hometown',
           5: 'birth_date',
           6: 'death_date',
           7: 'a_avg_rating',
           8: 'a_ratings_count',
           9: 'a_text_reviews_count'}
authors_df.rename(columns=columns, inplace=True)

len(authors_df)

with open('../data/raw/authors_scraped.pkl', 'wb') as picklefile:
    pickle.dump(authors_df, picklefile)

### Get info about book by book id

url = https://www.goodreads.com/book/show/1017661?key=2Kp7IGlO7GKjmAASOT2pEw

Requires
* book ID

Returns
* in book
    * publisher: Publisher
    * rating_dist: distribution of ratings

with open('../data/raw/books_scraped.pkl', 'rb') as picklefile:
    books_df = pickle.load(picklefile)

b_ids = books_df['b_id']

books_base_url = 'https://www.goodreads.com/book/show/'
search_criteria = {'key': key}

books_urls = {}

for b_id in b_ids:
    url = books_base_url + b_id + '?' + urllib.urlencode(search_criteria)
    books_urls[url] = b_id

def get_ratings_dist(rating_dist):
    for i in range(0, 5):
        yield rating_dist[i].split(':')[1]

def get_book_data(book, b_id, df):
    '''
    Retrieve data from books xml tree on author paginate page
    '''
    publisher = book.find('publisher').text
    rating_dist = book.find('work').find('rating_dist').text.split('|')
    ratings = get_ratings_dist(rating_dist)
    fives, fours, threes, twos, ones = get_ratings_dist(rating_dist)
    results = pd.Series((b_id, publisher, fives, fours, threes, twos, ones))
    return df.append(results, ignore_index=True)

#get data by book id from book info API
pubs_df = pd.DataFrame(data=None)
count = 0

for url, b_id in books_urls.items():
    soup = try_url(url)
    book = ET.fromstring(str(soup.find('book')))
    pubs_df = get_book_data(book, b_id, pubs_df)
    count += 1
    print b_id, count
    sleep(1.5)

columns = {0: 'b_id',
           1: 'publisher',
           2: 'b_rating_5_count',
           3: 'b_rating_4_count',
           4: 'b_rating_3_count',
           5: 'b_rating_2_count',
           6: 'b_rating_1_count',}

pubs_df.rename(columns=columns, inplace=True)

books_df = books_df.merge(pubs_df, on='b_id')

with open('../data/raw/books_scraped.pkl', 'wb') as picklefile:
    pickle.dump(books_df, picklefile)
