==> 02_luther/notebooks/04_goodreads_cleaning.ipynb <==
import pandas as pd
import numpy as np
import pickle

%matplotlib inline

#load pickled file into dataframe
def open_pickle(pkl_file):
    with open(pkl_file, 'rb') as picklefile:
        return pickle.load(picklefile)

#load data collected from wikipedia and goodreads
nyt = open_pickle('../data/interim/nyt_cleaned.pkl')
books = open_pickle('../data/raw/books_scraped.pkl')
authors = open_pickle('../data/raw/authors_scraped.pkl')

#convert number strings to numeric values
def to_num(df):
    return df.apply(lambda x: pd.to_numeric(x, errors='ignore'))

nyt = to_num(nyt)
books = to_num(books)
authors = to_num(authors)

### Clean author data

#add author ids to authors df
a_ids = books[['a_id', 'author']].drop_duplicates()
authors = authors.merge(a_ids, on='a_id')

#drop birth_date, death_date because of number of NANs
print authors.isnull().sum()/len(authors)
labels = ['birth_date', 'death_date']
authors.drop(labels=labels, inplace=True, axis=1)

#fill gender and hometown NANs with 'not reported'
authors['gender'].fillna('not reported', inplace=True)
authors['hometown'].fillna('not reported', inplace=True)

#calculate number of weeks each book was on NYT lists by book, and by author
nyt_by_book = nyt.groupby(by=['title', 'author'], as_index=False).count()[['title', 'author', 'date']]
nyt_by_author = nyt.groupby(by='author', as_index=False).count()[['author', 'date']]

#rename columns to be able to join nyt df with goodreads dfs
columns = {'title': 'nyt_title',
           'date': 'b_wks_on_list'}

nyt_by_book.rename(columns=columns, inplace=True)
nyt_by_author.rename(columns={'date': 'a_wks_on_list'}, inplace=True)

#calculate number of books on nyt for each author
nyt_books_by_author = nyt.groupby(['author','title'], as_index=False).count()[['author','title']]
nyt_books_by_author = nyt_books_by_author.groupby('author', as_index=False).count()
nyt_books_by_author.rename(columns={'title': 'a_books_on_list'}, inplace=True)

#add number of weeks each author has been on NYT list
nyt_by_author['author'] = nyt_by_author['author'].apply(lambda x: x.decode('utf-8'))
authors = authors.merge(nyt_by_author, on='author', how='left')

#add number of weeks each author has on NYT list
nyt_books_by_author['author'] = nyt_books_by_author['author'].apply(lambda x: x.decode('utf-8'))
authors = authors.merge(nyt_books_by_author, on='author', how='left')

authors.loc[authors['author'].str.contains('Betty'), 'a_wks_on_list'] = 22
authors.loc[authors['author'].str.contains('Betty'), 'a_books_on_list'] = 1

### Clean books data

#determine which decade each book was on the NYT list
#if in more than one decade, only keep first decade (about 6 books)
books_by_decade = nyt.groupby(['title', 'decade'], as_index=False).count()[['title', 'decade']]
books_by_decade.drop_duplicates(subset='title', inplace=True)

#add number of weeks each book was on NYT list
books = books.merge(nyt_by_book, on='nyt_title', how='left')

#add author data to each book
books = books.merge(authors, on='a_id')

#add which decade a book was in for each book
books_by_decade.rename(columns={'title': 'nyt_title'}, inplace=True)
books = books.merge(books_by_decade, on='nyt_title')

#drop author names from books df
books = books.drop(['author_x', 'author_y'], axis=1)

#fix typo in pub_yr
books.loc[books['pub_yr']==214,'pub_yr'] = 2014

#drop pub_day, pub_mon because of number of NANs
print books.isnull().sum()/len(books)
labels = ['pub_day', 'pub_mon']
books.drop(labels=labels, inplace=True, axis=1)

#fill publisher NANs with 'Unknown'
books['publisher'].fillna('Unknown', inplace=True)

#standardize publisher names
publishers = {'Berkley': 'Berkley',
              'Brown' : 'Little Brown & Co',
              'Bantam': 'Bantam',
              'Putnam': 'G.P. Putnam\'s',
              'Dell': 'Dell',
              'Pocket': 'Pocket',
              'Scribner': 'Scribner',
              'Doubleday': 'Doubleday',
              'Arrow': 'Arrow',
              'Atria': 'Atria',
              'Vintage': 'Vintage',
              'Random': 'Random House',
              'Simon': 'Simon & Schuster',
              'Signet': 'Signet',
              'Martin': 'St. Martin\'s Press',
              'Penguin': 'Penguin',
              'Warner': 'Warner',
              'NAL': 'NAL',
              'Orion': 'Orion',
              'Dial': 'Dial',
              'Corgi': 'Corgi',
              'Harper': 'Harper Collins',
              'Collins': 'Harper Collins',
              'Ace': 'Ace',
              'Knopf': 'Knopf',
              'Fawcett': 'Fawcett',
              'Tor': 'Tor',
              'New English': 'New English',
              'Headline': 'Headline',
              'Avon': 'Avon',
              'Hachette': 'Hachette',
              'Viking': 'Viking',
              'Morrow': 'Morrow',
              'Hodder': 'Hodder & Stoughton',
              'Picador': 'Picador',
              'Piatkus': 'Piatkus',
              'MacMillan': 'MacMillan',
              'Houghton': 'Houghton Mifflin',
              'B E': 'B E Trice',
              'Dodd': 'Dodd Mead'
             }

for old_pub, new_pub in publishers.items():
    books.loc[books['publisher'].str.contains(old_pub), 'publisher'] = new_pub

### Pickle cleaned books and authors dataframes to be used in analysis

books.to_pickle('../data/processed/books_cleaned.pkl')
authors.to_pickle('../data/processed/authors_cleaned.pkl')
